+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## UnRelease
 
+## 0.0.5
+
 - Add abbr for `--option-file`.
 - Fix cmake compile problem for x86_64.